+++ tencentcloud/cwp/v20180228/models.go
@@ -12348,6 +12348,9 @@ type DescribeMalwareTimingScanSettingResponseParams struct {
 	// 1标准模式（只报严重、高危）、2增强模式（报严重、高危、中危）、3严格模式（报严重、高、中、低、提示）
 	EngineType *uint64 `json:"EngineType,omitempty" name:"EngineType"`
 
+	// 启发引擎 0 关闭 1开启
+	EnableInspiredEngine *uint64 `json:"EnableInspiredEngine,omitempty" name:"EnableInspiredEngine"`
+
 	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
 	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
 }
@@ -19553,6 +19556,9 @@ type ModifyMalwareTimingScanSettingsRequestParams struct {
 
 	// 1标准模式（只报严重、高危）、2增强模式（报严重、高危、中危）、3严格模式（报严重、高、中、低、提示）
 	EngineType *uint64 `json:"EngineType,omitempty" name:"EngineType"`
+
+	// 启发引擎开关 0 关闭 1开启
+	EnableInspiredEngine *uint64 `json:"EnableInspiredEngine,omitempty" name:"EnableInspiredEngine"`
 }
 
 type ModifyMalwareTimingScanSettingsRequest struct {
@@ -19593,6 +19599,9 @@ type ModifyMalwareTimingScanSettingsRequest struct {
 
 	// 1标准模式（只报严重、高危）、2增强模式（报严重、高危、中危）、3严格模式（报严重、高、中、低、提示）
 	EngineType *uint64 `json:"EngineType,omitempty" name:"EngineType"`
+
+	// 启发引擎开关 0 关闭 1开启
+	EnableInspiredEngine *uint64 `json:"EnableInspiredEngine,omitempty" name:"EnableInspiredEngine"`
 }
 
 func (r *ModifyMalwareTimingScanSettingsRequest) ToJsonString() string {
@@ -19619,6 +19628,7 @@ func (r *ModifyMalwareTimingScanSettingsRequest) FromJsonString(s string) error 
 	delete(f, "AutoIsolation")
 	delete(f, "KillProcess")
 	delete(f, "EngineType")
+	delete(f, "EnableInspiredEngine")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "ModifyMalwareTimingScanSettingsRequest has unknown keys!", "")
 	}