+++ tencentcloud/ivld/v20210903/models.go
@@ -1922,6 +1922,10 @@ type PersonInfo struct {
 
 	// 人物出现信息
 	AppearInfo *AppearInfo `json:"AppearInfo,omitempty" name:"AppearInfo"`
+
+	// 人脸在图片中的位置，仅在图片标签任务有效
+	// 注意：此字段可能返回 null，表示取不到有效值。
+	AppearRect *Rectf `json:"AppearRect,omitempty" name:"AppearRect"`
 }
 
 // Predefined struct for user