+++ tencentcloud/live/v20180801/client.go
@@ -2930,7 +2930,7 @@ func NewDeletePullStreamConfigResponse() (response *DeletePullStreamConfigRespon
 }
 
 // DeletePullStreamConfig
-// 删除直播拉流配置。
+// 删除直播拉流配置。该接口已下线,请使用新接口 DeleteLivePullStreamTask。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -2949,7 +2949,7 @@ func (c *Client) DeletePullStreamConfig(request *DeletePullStreamConfigRequest) 
 }
 
 // DeletePullStreamConfig
-// 删除直播拉流配置。
+// 删除直播拉流配置。该接口已下线,请使用新接口 DeleteLivePullStreamTask。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -6718,7 +6718,7 @@ func NewDescribePullStreamConfigsResponse() (response *DescribePullStreamConfigs
 }
 
 // DescribePullStreamConfigs
-// 查询直播拉流配置。
+// 查询直播拉流配置。该接口已下线,请使用新接口 DescribeLivePullStreamTasks。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -6739,7 +6739,7 @@ func (c *Client) DescribePullStreamConfigs(request *DescribePullStreamConfigsReq
 }
 
 // DescribePullStreamConfigs
-// 查询直播拉流配置。
+// 查询直播拉流配置。该接口已下线,请使用新接口 DescribeLivePullStreamTasks。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -8660,7 +8660,7 @@ func NewModifyPullStreamConfigResponse() (response *ModifyPullStreamConfigRespon
 }
 
 // ModifyPullStreamConfig
-// 更新拉流配置。
+// 更新拉流配置。该接口为已下线接口，请使用新接口 ModifyLivePullStreamTask。 
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -8681,7 +8681,7 @@ func (c *Client) ModifyPullStreamConfig(request *ModifyPullStreamConfigRequest) 
 }
 
 // ModifyPullStreamConfig
-// 更新拉流配置。
+// 更新拉流配置。该接口为已下线接口，请使用新接口 ModifyLivePullStreamTask。 
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -8732,7 +8732,7 @@ func NewModifyPullStreamStatusResponse() (response *ModifyPullStreamStatusRespon
 }
 
 // ModifyPullStreamStatus
-// 修改直播拉流配置的状态。
+// 修改直播拉流配置的状态。该接口已下线,请使用新接口 ModifyLivePullStreamTask。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"
@@ -8752,7 +8752,7 @@ func (c *Client) ModifyPullStreamStatus(request *ModifyPullStreamStatusRequest) 
 }
 
 // ModifyPullStreamStatus
-// 修改直播拉流配置的状态。
+// 修改直播拉流配置的状态。该接口已下线,请使用新接口 ModifyLivePullStreamTask。
 //
 // 可能返回的错误码:
 //  FAILEDOPERATION_CALLOTHERSVRFAILED = "FailedOperation.CallOtherSvrFailed"