+++ tencentcloud/tag/v20180813/errors.go
@@ -26,6 +26,12 @@ const (
 	// 操作失败。
 	FAILEDOPERATION = "FailedOperation"
 
+	// 项目已禁用。
+	FAILEDOPERATION_PROJECTDISABLED = "FailedOperation.ProjectDisabled"
+
+	// 项目数超过限制。
+	FAILEDOPERATION_PROJECTNUMEXCEED = "FailedOperation.ProjectNumExceed"
+
 	// 单次请求的资源appId必须相同。
 	FAILEDOPERATION_RESOURCEAPPIDNOTSAME = "FailedOperation.ResourceAppIdNotSame"
 
@@ -43,6 +49,9 @@ const (
 
 	// PaginationToken参数非法。
 	INVALIDPARAMETER_PAGINATIONTOKENINVALID = "InvalidParameter.PaginationTokenInvalid"
+
+	// 项目名称已存在。
+	INVALIDPARAMETER_PROJECTNAMEEXISTED = "InvalidParameter.ProjectNameExisted"
 
 	// 系统预留标签键 qcloud、tencent和project 禁止创建。
 	INVALIDPARAMETER_RESERVEDTAGKEY = "InvalidParameter.ReservedTagKey"
@@ -142,6 +151,9 @@ const (
 
 	// 资源关联的标签键不存在。
 	RESOURCENOTFOUND_ATTACHEDTAGKEYNOTFOUND = "ResourceNotFound.AttachedTagKeyNotFound"
+
+	// 记录不存在。
+	RESOURCENOTFOUND_RECORDNOTFOUND = "ResourceNotFound.RecordNotFound"
 
 	// 标签不存在。
 	RESOURCENOTFOUND_TAGNONEXIST = "ResourceNotFound.TagNonExist"