+++ tencentcloud/tag/v20180813/models.go
@@ -21,6 +21,73 @@ import (
 )
 
 // Predefined struct for user
+type AddProjectRequestParams struct {
+	// 项目名称
+	ProjectName *string `json:"ProjectName,omitempty" name:"ProjectName"`
+
+	// 项目描述
+	Info *string `json:"Info,omitempty" name:"Info"`
+}
+
+type AddProjectRequest struct {
+	*tchttp.BaseRequest
+	
+	// 项目名称
+	ProjectName *string `json:"ProjectName,omitempty" name:"ProjectName"`
+
+	// 项目描述
+	Info *string `json:"Info,omitempty" name:"Info"`
+}
+
+func (r *AddProjectRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *AddProjectRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	delete(f, "ProjectName")
+	delete(f, "Info")
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "AddProjectRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type AddProjectResponseParams struct {
+	// 项目Id
+	ProjectId *uint64 `json:"ProjectId,omitempty" name:"ProjectId"`
+
+	// 是否为新项目
+	IsNew *int64 `json:"IsNew,omitempty" name:"IsNew"`
+
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type AddProjectResponse struct {
+	*tchttp.BaseResponse
+	Response *AddProjectResponseParams `json:"Response"`
+}
+
+func (r *AddProjectResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *AddProjectResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
 type AddResourceTagRequestParams struct {
 	// 标签键
 	TagKey *string `json:"TagKey,omitempty" name:"TagKey"`
@@ -2520,6 +2587,81 @@ func (r *UnTagResourcesResponse) ToJsonString() string {
 // FromJsonString It is highly **NOT** recommended to use this function
 // because it has no param check, nor strict type check
 func (r *UnTagResourcesResponse) FromJsonString(s string) error {
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type UpdateProjectRequestParams struct {
+	// 项目ID
+	ProjectId *uint64 `json:"ProjectId,omitempty" name:"ProjectId"`
+
+	// 项目名称
+	ProjectName *string `json:"ProjectName,omitempty" name:"ProjectName"`
+
+	// 禁用项目，1，禁用，0，启用
+	Disable *int64 `json:"Disable,omitempty" name:"Disable"`
+
+	// 备注
+	Info *string `json:"Info,omitempty" name:"Info"`
+}
+
+type UpdateProjectRequest struct {
+	*tchttp.BaseRequest
+	
+	// 项目ID
+	ProjectId *uint64 `json:"ProjectId,omitempty" name:"ProjectId"`
+
+	// 项目名称
+	ProjectName *string `json:"ProjectName,omitempty" name:"ProjectName"`
+
+	// 禁用项目，1，禁用，0，启用
+	Disable *int64 `json:"Disable,omitempty" name:"Disable"`
+
+	// 备注
+	Info *string `json:"Info,omitempty" name:"Info"`
+}
+
+func (r *UpdateProjectRequest) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *UpdateProjectRequest) FromJsonString(s string) error {
+	f := make(map[string]interface{})
+	if err := json.Unmarshal([]byte(s), &f); err != nil {
+		return err
+	}
+	delete(f, "ProjectId")
+	delete(f, "ProjectName")
+	delete(f, "Disable")
+	delete(f, "Info")
+	if len(f) > 0 {
+		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "UpdateProjectRequest has unknown keys!", "")
+	}
+	return json.Unmarshal([]byte(s), &r)
+}
+
+// Predefined struct for user
+type UpdateProjectResponseParams struct {
+	// 唯一请求 ID，每次请求都会返回。定位问题时需要提供该次请求的 RequestId。
+	RequestId *string `json:"RequestId,omitempty" name:"RequestId"`
+}
+
+type UpdateProjectResponse struct {
+	*tchttp.BaseResponse
+	Response *UpdateProjectResponseParams `json:"Response"`
+}
+
+func (r *UpdateProjectResponse) ToJsonString() string {
+    b, _ := json.Marshal(r)
+    return string(b)
+}
+
+// FromJsonString It is highly **NOT** recommended to use this function
+// because it has no param check, nor strict type check
+func (r *UpdateProjectResponse) FromJsonString(s string) error {
 	return json.Unmarshal([]byte(s), &r)
 }