+++ CHANGELOG.md
@@ -1,3 +1,258 @@
+# Release v1.0.522
+
+## DDoS 防护(antiddos) 版本：2020-03-09
+
+### 第 43 次发布
+
+发布时间：2022-10-26 06:07:49
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DescribeListBGPInstances](https://cloud.tencent.com/document/api/1021/57521)
+
+	* 新增入参：FilterLightFlag, FilterChannelFlag
+
+
+
+
+## 云呼叫中心(ccc) 版本：2020-02-10
+
+### 第 32 次发布
+
+发布时间：2022-10-26 06:15:12
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [CreateAutoCalloutTask](https://cloud.tencent.com/document/api/679/69194)
+
+	* 新增入参：Variables
+
+
+新增数据结构：
+
+* [Variable](https://cloud.tencent.com/document/api/679/47715#Variable)
+
+
+
+## 日志服务(cls) 版本：2020-10-16
+
+### 第 31 次发布
+
+发布时间：2022-10-26 06:21:36
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [SearchLog](https://cloud.tencent.com/document/api/614/56447)
+
+	* 新增入参：SamplingRate
+
+
+
+
+## 企业收付平台(cpdp) 版本：2019-08-20
+
+### 第 92 次发布
+
+发布时间：2022-10-26 06:23:03
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [UnifiedCloudOrder](https://cloud.tencent.com/document/api/1122/72599)
+
+	* 新增入参：ExternalAttachmentDataList
+
+
+新增数据结构：
+
+* [CloudExternalAttachmentData](https://cloud.tencent.com/document/api/1122/40683#CloudExternalAttachmentData)
+
+修改数据结构：
+
+* [CloudSubOrder](https://cloud.tencent.com/document/api/1122/40683#CloudSubOrder)
+
+	* 新增成员：ExternalAttachmentDataList
+
+
+
+
+## 主机安全(cwp) 版本：2018-02-28
+
+### 第 60 次发布
+
+发布时间：2022-10-26 06:25:43
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改接口：
+
+* [DescribeMalwareTimingScanSetting](https://cloud.tencent.com/document/api/296/58240)
+
+	* 新增出参：EnableInspiredEngine
+
+* [ModifyMalwareTimingScanSettings](https://cloud.tencent.com/document/api/296/52509)
+
+	* 新增入参：EnableInspiredEngine
+
+
+
+
+## 电子签服务(ess) 版本：2020-11-11
+
+### 第 21 次发布
+
+发布时间：2022-10-26 06:35:22
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增数据结构：
+
+* [ApproverOption](https://cloud.tencent.com/document/api/1323/70369#ApproverOption)
+
+修改数据结构：
+
+* [FlowCreateApprover](https://cloud.tencent.com/document/api/1323/70369#FlowCreateApprover)
+
+	* 新增成员：ApproverOption
+
+
+
+
+## 腾讯电子签（基础版）(essbasic) 版本：2021-05-26
+
+### 第 38 次发布
+
+发布时间：2022-10-26 06:35:40
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增数据结构：
+
+* [ApproverOption](https://cloud.tencent.com/document/api/1420/61525#ApproverOption)
+
+修改数据结构：
+
+* [FlowApproverInfo](https://cloud.tencent.com/document/api/1420/61525#FlowApproverInfo)
+
+	* 新增成员：ApproverOption, ApproverNeedSignReview
+
+
+
+
+## 腾讯电子签（基础版）(essbasic) 版本：2020-12-22
+
+
+
+## 媒体智能标签(ivld) 版本：2021-09-03
+
+### 第 8 次发布
+
+发布时间：2022-10-25 19:11:18
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [PersonInfo](https://cloud.tencent.com/document/api/1509/65063#PersonInfo)
+
+	* 新增成员：AppearRect
+
+
+
+
+## 云数据库 MariaDB(mariadb) 版本：2017-03-12
+
+### 第 43 次发布
+
+发布时间：2022-10-26 06:46:25
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [ModifyInstanceNetwork](https://cloud.tencent.com/document/api/237/81949)
+* [ModifyInstanceVip](https://cloud.tencent.com/document/api/237/81948)
+* [ModifyInstanceVport](https://cloud.tencent.com/document/api/237/81947)
+
+
+
+## 私有域解析 Private DNS(privatedns) 版本：2020-10-28
+
+### 第 11 次发布
+
+发布时间：2022-10-26 06:52:45
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [ModifyRecordsStatus](https://cloud.tencent.com/document/api/1338/81950)
+
+
+
+## 邮件推送(ses) 版本：2020-10-02
+
+### 第 15 次发布
+
+发布时间：2022-10-26 06:55:19
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+修改数据结构：
+
+* [CycleEmailParam](https://cloud.tencent.com/document/api/1288/51053#CycleEmailParam)
+
+	* 新增成员：TermCycle
+
+
+
+
+## 标签(tag) 版本：2018-08-13
+
+### 第 12 次发布
+
+发布时间：2022-10-26 06:58:19
+
+本次发布包含了以下内容：
+
+改善已有的文档。
+
+新增接口：
+
+* [AddProject](https://cloud.tencent.com/document/api/651/81952)
+* [UpdateProject](https://cloud.tencent.com/document/api/651/81951)
+
+
+
 # Release v1.0.521
 
 ## 内容分发网络(cdn) 版本：2018-06-06